fixed path launcher delete

+++ rofi-path-launcher.sh
@@ -27,13 +27,20 @@ if [[ "$config" == "" ]]; then
 		touch $config_path
 		echo "{}" > $config_path
 	fi
+	
+	config=$(cat $config_path)
+
+	if [[ "$config" == "" ]]; then
+		echo "{}" > $config_path
+	fi
+
 	config=$(cat $config_path)
 fi
 
 function get_program {
 	programs_num=$({ echo $config | jq '. |= keys' | jq -r -c '.[]'; echo $ADD_OPTION; echo $REMOVE_OPTION; } | wc -l) 
 	if [[ $programs_num -gt 10 ]]; then programs_num=10; fi
-	program=$({ echo $config | jq '. |= keys' | jq -r -c '.[]'; echo $ADD_OPTION; echo $REMOVE_OPTION; } | rofi -dmenu -i -theme $theme -l $programs_num -p " Path Launcher:")
+	program=$({ echo $config | jq '. |= keys' | jq -r -c '.[]' | sort; echo $ADD_OPTION; echo $REMOVE_OPTION; } | rofi -dmenu -i -theme $theme -l $programs_num -p " Path Launcher:")
 	if [[ $program == $ADD_OPTION ]]; then add_program;
 	elif [[ $program == $REMOVE_OPTION ]]; then remove_program;
 	elif [[ $program != "" ]]; then get_program_path $program; 
@@ -44,7 +51,7 @@ function get_program_path {
 	program="$1"
 	
 	pattern=".${program}.paths[]"
-	lines_num=$({ echo $config | jq -r -c $pattern | sort; echo $ADD_OPTION; echo $REMOVE_OPTION; } | wc -l)
+	lines_num=$({ echo $config | jq -r -c $pattern; echo $ADD_OPTION; echo $REMOVE_OPTION; } | wc -l)
 	if [[ $lines_num -gt 10 ]]; then lines_num=10; fi
 	path=$({ echo $config | jq -r -c $pattern | sort; echo $ADD_OPTION; echo $REMOVE_OPTION; }  | rofi -dmenu -i -theme $theme -l $lines_num -p " $program (path)")
 
@@ -69,7 +76,7 @@ function execute_command {
 function add_path_to_program {
 	program="$1"
 
-	path=$(echo "" | rofi -dmenu -i -theme $theme -l 0 -p " $program (add path)")
+	path=$(rofi -dmenu -i -theme $theme -l 0 -p " $program (add path)")
 
 	if [[ "$path" == "" ]]
 		then get_program_path $program
@@ -98,47 +105,54 @@ function remove_path_from_program {
 	path=$(echo $config | jq -r -c $pattern | sort | rofi -dmenu -i -theme $theme -l $lines_num -p " $program (remove path)")
 
 	if [[ "$path" != "" ]]; then
-		pattern=".${program}.paths - [\"${path}\"]"
-		new_paths=$(jq "$pattern" $config_path)
+		confirm=$(ask_for_confirmation "$path")
+		if [[ ${confirm,,} == "yes" ]]; then 
+			pattern=".${program}.paths - [\"${path}\"]"
+			new_paths=$(jq "$pattern" $config_path)
 
-		pattern=".${program}.paths = ${new_paths}"
-		new_config=$(jq "$pattern" $config_path)
+			pattern=".${program}.paths = ${new_paths}"
+			new_config=$(jq "$pattern" $config_path)
 
-		echo $new_config > $config_path
-		config=$(cat $config_path)
+			echo $new_config > $config_path
+			config=$(cat $config_path)
+		fi
 	fi
 
 	get_program_path $1
 }
 
 function add_program {
-	program_name=$(echo "" | rofi -dmenu -i -theme $theme -l 0 -p " Path Launcher (program name)")
+	program_name=$(rofi -dmenu -i -theme $theme -l 0 -p " Path Launcher (program name)")
 
-	if [[ "$program_name" == "" ]]; then get_program;
-	else 
-		program_command=$(echo "" | rofi -dmenu -i -theme $theme -l 0 -p " Path Launcher (program cmd)")
+	if [[ "$program_name" == "" ]]; 
+		then get_program;
+		else 
+			program_command=$(rofi -dmenu -i -theme $theme -l 0 -p " Path Launcher (program cmd)")
 
-		pattern=". + {\"${program_name}\":{\"command\":\"${program_command}\",\"paths\":[]}}"
-		new_config=$(jq "$pattern" $config_path)
+			pattern=". + {\"${program_name}\":{\"command\":\"${program_command}\",\"paths\":[]}}"
+			new_config=$(jq "$pattern" $config_path)
 
-		echo $new_config > $config_path
-		config=$(cat $config_path)
+			echo $new_config > $config_path
+			config=$(cat $config_path)
 
-		get_program;
-	fi
+			get_program;
+		fi
 }
 
 function remove_program {
 	programs_num=$(echo $config | jq '. |= keys' | jq -r -c '.[]' | wc -l) 
 	if [[ $programs_num -gt 10 ]]; then programs_num=10; fi
-	program=$(echo $config | jq '. |= keys' | jq -r -c '.[]' | rofi -dmenu -i -theme $theme -l $programs_num -p " Path Launcher:")
+	program=$(echo $config | jq '. |= keys' | jq -r -c '.[]' | sort | rofi -dmenu -i -theme $theme -l $programs_num -p " Path Launcher (remove program):")
 
 	if [[ "$program" != "" ]]; then 
-		pattern="del(.$program)"
-		new_config=$(jq "$pattern" $config_path)
-		
-		echo $new_config > $config_path
-		config=$(cat $config_path)
+		confirm=$(ask_for_confirmation "$program")
+		if [[ ${confirm,,} == "yes" ]]; then 
+			pattern="del(.$program)"
+			new_config=$(jq "$pattern" $config_path)
+			
+			echo $new_config > $config_path
+			config=$(cat $config_path)
+		fi
 	fi
 
 	get_program
@@ -146,6 +160,10 @@ function remove_program {
 
 function print_error {
 	rofi -e "Error: $1" 
+}
+
+function ask_for_confirmation {
+	echo $(rofi -dmenu -theme $theme -p " Confirm ($1) ? [yes/no]" -l 0)
 }
 
 get_program
